fix: web client build
