+++ crates/web-client/js/index.js
@@ -10,6 +10,7 @@ const {
   Felt,
   FeltArray,
   FungibleAsset,
+  InputNoteState,
   Note,
   NoteAssets,
   NoteExecutionHint,
@@ -22,7 +23,6 @@ const {
   NoteInputs,
   NoteMetadata,
   NoteRecipient,
-  NoteState,
   NoteTag,
   NoteType,
   OutputNote,
@@ -50,6 +50,7 @@ export {
   Felt,
   FeltArray,
   FungibleAsset,
+  InputNoteState,
   Note,
   NoteAssets,
   NoteExecutionHint,
@@ -62,7 +63,6 @@ export {
   NoteInputs,
   NoteMetadata,
   NoteRecipient,
-  NoteState,
   NoteTag,
   NoteType,
   OutputNote,
