+++ crates/web-client/js/types/index.d.ts
@@ -8,6 +8,7 @@ export {
   Felt,
   FeltArray,
   FungibleAsset,
+  InputNoteState,
   NewSwapTransactionResult,
   NewTransactionResult,
   Note,
@@ -22,7 +23,6 @@ export {
   NoteInputs,
   NoteMetadata,
   NoteRecipient,
-  NoteState,
   NoteTag,
   NoteType,
   OutputNote,
