+++ crates/web-client/src/models/mod.rs
@@ -43,6 +43,7 @@ pub mod felt;
 pub mod fungible_asset;
 pub mod input_note;
 pub mod input_note_record;
+pub mod input_note_state;
 pub mod input_notes;
 pub mod merkle_path;
 pub mod note;
@@ -59,7 +60,6 @@ pub mod note_location;
 pub mod note_metadata;
 pub mod note_recipient;
 pub mod note_script;
-pub mod note_state;
 pub mod note_tag;
 pub mod note_type;
 pub mod output_note;
