+++ crates/web-client/test.html
@@ -23,12 +23,12 @@
             Felt,
             FeltArray,
             FungibleAsset,
+            InputNoteState,
             Note,
             NoteAssets,
             NoteExecutionHint,
             NoteExecutionMode,
             NoteFilter,
-            NoteState,
             NoteId,
             NoteIdAndArgs,
             NoteIdAndArgsArray,
